changes to align with rsbc/7.0.0
